+++ apps/backoffice-v2/src/domains/business-reports/fetchers.ts
@@ -2,21 +2,18 @@ import qs from 'qs';
 import { z } from 'zod';
 import { t } from 'i18next';
 import { toast } from 'sonner';
-import { UnknownRecord } from 'type-fest';
 
 import { Method } from '@/common/enums';
 import { apiClient } from '@/common/api-client/api-client';
 import { TReportStatusValue, TRiskLevel } from '@/pages/MerchantMonitoring/schemas';
 import { handleZodError } from '@/common/utils/handle-zod-error/handle-zod-error';
 import {
-  MERCHANT_REPORT_STATUSES,
   MERCHANT_REPORT_STATUSES_MAP,
-  MERCHANT_REPORT_TYPES,
-  MERCHANT_REPORT_VERSIONS,
   MerchantReportStatus,
   MerchantReportType,
   MerchantReportVersion,
   ReportSchema,
+  UPDATEABLE_REPORT_STATUSES,
 } from '@ballerine/common';
 
 const statusOverrides = {
@@ -24,13 +21,17 @@ const statusOverrides = {
   [MERCHANT_REPORT_STATUSES_MAP['quality-control']]: MERCHANT_REPORT_STATUSES_MAP['in-progress'],
 } as const satisfies Partial<Record<MerchantReportStatus, MerchantReportStatus>>;
 
-export const BusinessReportSchema = ReportSchema.transform(data => ({
-  ...data,
-  status: data.status in statusOverrides ? statusOverrides[data.status] : data.status,
-  website: data.website.url,
-  riskLevel: data.status === MERCHANT_REPORT_STATUSES_MAP.completed ? data.riskLevel : null,
-  data: data.status === MERCHANT_REPORT_STATUSES_MAP.completed ? data?.data : null,
-}));
+export const BusinessReportSchema = ReportSchema.transform(data => {
+  const isReportReady = UPDATEABLE_REPORT_STATUSES.includes(data.status);
+
+  return {
+    ...data,
+    status: data.status in statusOverrides ? statusOverrides[data.status] : data.status,
+    website: data.website.url,
+    riskLevel: isReportReady ? data.riskLevel : null,
+    data: isReportReady ? data?.data : null,
+  };
+});
 
 export const BusinessReportsSchema = z.object({
   data: z.array(BusinessReportSchema),