+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -140,7 +140,7 @@ export const useDefaultBlocksLogic = () => {
       omitPropsFromObjectWhitelist({
         object: workflow?.context?.pluginsOutput,
         whitelist: registryInfoWhitelist,
-      }),
+      }) ?? {},
     [workflow?.context?.pluginsOutput],
   );