+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.97
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.77
+  - @ballerine/workflow-core@0.6.97
+
+## 0.6.96
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.76
+  - @ballerine/workflow-core@0.6.96
+
 ## 0.6.95
 
 ### Patch Changes