+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.97
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.97
+
+## 0.6.96
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.96
+
 ## 0.6.95
 
 ### Patch Changes